requirements.txt
torch>=2.0
torchvision>=0.15
opencv-python>=4.8
Pillow>=9.0
joblib>=1.3
pandas>=1.5
flask>=2.2
scikit-learn>=1.2
